+++ CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   moved to `wal-key-bindings`.
 - All packages now explicitly require the packages they depend on.
   Only relevant if `wal-packages` is altered before bootstrapping.
+- `wal-agenda-buffer-p` now uses `org-agenda-file-p`.
 
 ## [v2.0.3]